pyproject.toml
[tool.poetry]
name = "PyKantorovich"
version = "0.1.0"
description = "Kantorovich distance between probabilities on a finite space."
authors = ["Stéphane Laurent <laurent_step@outlook.fr>"]
license = "GPL-3.0-only"

[tool.poetry.dependencies]
python = ">=3.7.1,<3.10"
numpy = "^1.21.2"
pycddlib = "^2.1.4"
scipy = "^1.7.1"
cvxpy = "^1.1.17"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
